+++ library/src/main/jni/com_twilio_video_MediaFactory.cpp
@@ -244,8 +244,9 @@ JNIEXPORT jlong JNICALL Java_com_twilio_video_MediaFactory_nativeCreate(JNIEnv *
     webrtc_jni::MediaCodecVideoDecoderFactory* video_decoder_factory =
             new webrtc_jni::MediaCodecVideoDecoderFactory();
 
-    // Enable use of textures for encoding and decoding
-    video_encoder_factory->SetEGLContext(jni, j_egl_local_context);
+    // Disabling capturing and encoding to texture until we understand GSDK-1132 GSDK-1139
+    // video_encoder_factory->SetEGLContext(jni, j_egl_local_context);
+    // Enable use of textures for decoding
     video_decoder_factory->SetEGLContext(jni, j_egl_remote_context);
 
     media_options.video_encoder_factory = video_encoder_factory;